Rewrite to deal with nested Tweet attributes
To allow parsing of nested Tweet attributes, e.g. user.id,
create the classes NestedDict and Tweet (a subclass of
NestedTweet). Since Tweet validation is now done by the Tweet
object itself, remove the TweetValidator class.

Include a module for testing NestedDict and Tweet classes.

Make the logic of the tweets_to_dataframe method of the
ParseTweets class cleaner and easier to follow.

Improve documentation and comments throughout.

This is effectively a rewrite of the module.